[publishWorkspace] - Resolving de-synchronization on multiple send requests (sub-tasks) in a one task (result to fast to handle response because of desynchronization)
probleme : CompletableFuture + quasar not working in this context with vertx. future.get() not allowed and not working too (it doesn't block the thread because, is not executed on the loop event vertx thread)

Notes : quasar fails + VertCompletableFuture fails + promises vertx fails

Solution used : Changing implementation with TaskExecutor. Creation of a custom task implementation. This allows to handle multiple tasks in an asynchronous mode.

+ Creation of TaskVerticleContext
+ adding TaskContext
+ minor modifications